moved test login to its own page
