+++ public/index.html
@@ -15,7 +15,6 @@
     <!-- Inserts the navbar into the page -->
     <nav class="navbar"></nav>
     <button onclick="iloveClicking()">Cool Final Project Website</button>
-    <a href="/api/users/auth/github">Login</a>
   </body>
 
 </html>
