+++ public/js/main.js
@@ -3,26 +3,6 @@
 function iloveClicking() {
 	alert("I LOVE CLICKING")
 }
-async function getLoggedInUser(){
-  const response = await fetch('/api/users/git/userName', {
-    method: 'get'
-  })
-  console.log("fired")
-  let pUser = document.getElementById("pUser")
-  if(response.ok){
-    pUser.innerText= "User: " + JSON.parse(await response.text())["userName"]
-  }
-  else{
-    pUser.innerText= ""
-  }
-}
-
-async function logout(){
-  const response = await fetch('/api/users/git/logout', {
-    method: 'POST'
-  })
-  window.location.href="/"
-}
 
 //Adds the navigation bar to the location of the navbar classon each page.
 //To add the nav bar to a page just add the following HTML element <nav class="navbar"></nav>
@@ -40,6 +20,5 @@ const nav = function () {
 }
 
 window.onload = function () {
-    getLoggedInUser().then()
     nav()
 }
