+++ public/testPages/testAuth.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
-    <script src="js/main.js"></script>
+    <script src="testAuth.js"></script>
 </head>
 <body>
