+++ public/testPages/testLogin.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+    <script src="testAuth.js"></script>
+</head>
+<body>
+<a href="/api/users/auth/github">Login</a>
+</body>
+</html>
